requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
py-mon==2.1.0
watchdog==6.0.0
Werkzeug==3.1.3
requests==2.31.0
python-dotenv==1.0.1
pymongo==4.15.3
